app/requirements.txt
psycopg2-binary==2.8.6
Flask==2.0.1
Jinja2==3.0.1